requirements.txt
discord.py
pytz
flask
aiosqlite